Adds in felinid limbs (Tail + ears)
