+++ GJ2022/Entities/Pawns/Health/Bodyparts/Organs/Organ.cs
@@ -24,6 +24,32 @@ namespace GJ2022.Entities.Pawns.Health.Bodyparts.Organs
             {
                 body.processingOrgans.Add(this);
             }
+            //Insert
+            Insert(body);
+        }
+
+        public bool Insert(Body body)
+        {
+            //Update body
+            body.Conciousness += ConciousnessFactor;
+            body.Manipulation += ManipulationFactor;
+            body.Movement += MovementFactor;
+            body.Hearing += HearingFactor;
+            body.Vision += VisionFactor;
+            //Overlay handling
+            if (body.Parent.UsesLimbOverlays)
+                AddOverlay(body.Parent.Renderable);
+            return true;
+        }
+
+        public override bool Remove()
+        {
+            if (!base.Remove())
+                return false;
+            //We can be garbage collected at this point
+            //TODO: Drop an organ item
+            Body = null;
+            return true;
         }
 
         public override void OnDestruction()
