+++ GJ2022/Entities/Pawns/Mobs/Humans/BionicHuman.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using GJ2022.Entities.Pawns.Health.Bodies;
-using GJ2022.Entities.Pawns.Health.Bodyparts.Limbs.Human.Bionic;
+using GJ2022.Entities.Pawns.Health.Bodyparts.Limbs.Limbs.Bionic;
 using GJ2022.Utility.MathConstructs;
 
 namespace GJ2022.Entities.Pawns.Mobs.Humans
