+++ GJ2022/GJ2022.csproj
@@ -122,6 +122,9 @@
     <None Update="Data\IconFiles\human_face.bmp">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
+    <None Update="Data\IconFiles\mutant_bodyparts.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
     <None Update="Data\IconFiles\text.bmp">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
